+++ R/learner_prioritylasso_classif_priority_lasso.R
@@ -1,84 +1,91 @@
 #' @title Classification Priority Lasso Learner
 #' @author HarutyunyanLiana
+#' @author bblodfon
 #' @name mlr_learners_classif.priority_lasso
 #'
 #' @description
-#' Patient outcome prediction based on multi-omics data taking practitioners’ preferences into account.
+#' Patient outcome prediction based on multi-omics data, taking practitioners'
+#' preferences into account.
 #' Calls `prioritylasso::prioritylasso()` from \CRANpkg{prioritylasso}.
+#' Only binary classification is supported.
+#'
+#' @section Scope and supported arguments:
+#' This learner intentionally exposes a focused subset of training and prediction arguments.
+#' It is designed to work well out of the box, without requiring extensive parameter tuning.
+#' Some arguments from `cv.glmnet()`, `glmnet()`, and `predict.prioritylasso()` are not included,
+#' because they are not consistently supported or forwarded through the full train/predict path
+#' (e.g. handling missing test data, or performing relaxed lasso fits).
+#'
+#' Please open an issue if there is a need for supporting more learner parameters.
 #'
 #' @section Initial parameter values:
 #' - `family` is set to `"binomial"` and cannot be changed
+#'
+#' @section Custom mlr3 parameters:
+#' - `adaptive.order`: if `TRUE`, the priority order of blocks is estimated from the data
+#' following Herrmann et al. (2021), instead of using the user-supplied block order.
+#' For each block, a Ridge regression (`alpha = 0`) is fit using `cv.glmnet()` on that block alone.
+#' The importance of a block is measured by the mean absolute coefficient (MAC) score at the
+#' `lambda.min` value (the lambda giving minimum cross-validation error).
+#' A penalty factor of `1 / MAC` is then assigned to each block.
+#' Blocks are sorted by increasing penalty factor, i.e., blocks with larger MAC
+#' (stronger average signal) receive higher priority (come first).
+#' Also, the block‑wise penalty factors are attached to the fitted model object as
+#' `learner$model$block.penalty.factors`.
+#'
+#' This method is useful when no domain knowledge is available to specify block priority.
+#' In this step, data are **standardized by default** (`standardize = TRUE`), but this can
+#' be overridden by the learner's `standardize` parameter. `lambda.min` is always used
+#' to derive the block priority.
+#' Additional arguments such as `nfolds`, `type.measure`, `weights` and `cox.ties` (if provided) are
+#' forwarded to each block‑wise `cv.glmnet()` fit.
+#' The `max.coef` parameter, if supplied, it is re‑ordered accordingly to align
+#' with the new block order.
+#'
+#' This parameter is ignored when fewer than two blocks are provided.
+#' It defaults to `FALSE` for backward compatibility.
 #'
 #' @templateVar id classif.priority_lasso
 #' @template learner
 #'
 #' @references
-#' `r format_bib("klau2018priolasso")`
+#' `r format_bib("klau2018priolasso", "herrmann_2021")`
 #'
 #' @template seealso_learner
 #' @template example_prioritylasso
 #' @export
-LearnerClassifPriorityLasso = R6Class("LearnerClassifPriorityLasso",
+LearnerClassifPriorityLasso = R6Class(
+  "LearnerClassifPriorityLasso",
   inherit = LearnerClassif,
   public = list(
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
+      # fmt: skip
       param_set = ps(
-        blocks               = p_uty(tags = c("train", "required")),
-        type.measure         = p_fct(levels = c("class", "auc"), tags = c("train", "required")),
-        max.coef             = p_uty(default = NULL, tags = "train"),
-        block1.penalization  = p_lgl(default = TRUE, tags = "train"),
-        lambda.type          = p_fct(default = "lambda.min", levels = c("lambda.min", "lambda.1se"), tags = c("train", "predict")), #nolint
-        standardize          = p_lgl(default = TRUE, tags = "train"),
-        nfolds               = p_int(default = 5L, lower = 1L, tags = "train"),
-        foldid               = p_uty(default = NULL, tags = "train"),
-        cvoffset             = p_lgl(default = FALSE, tags = "train"),
-        cvoffsetnfolds       = p_int(default = 10, lower = 1L, tags = "train"),
-        return.x             = p_lgl(default = TRUE, tags = "train"),
-        handle.missingtestdata = p_fct(c("none", "omit.prediction", "set.zero", "impute.block"), tags = "predict"),
+        # prioritylasso::prioritylasso() parameters
+        type.measure          = p_fct(levels = c("class", "auc"), tags = c("train", "required")),
+        blocks                = p_uty(tags = c("train", "required")),
+        max.coef              = p_uty(default = NULL, tags = "train"),
+        block1.penalization   = p_lgl(default = TRUE, tags = "train"),
+        lambda.type           = p_fct(default = "lambda.min", levels = c("lambda.min", "lambda.1se"), tags = "train"),
+        standardize           = p_lgl(default = TRUE, tags = "train"),
+        nfolds                = p_int(3L, default = 10L, tags = "train"),
+        foldid                = p_uty(default = NULL, tags = "train"),
+        cvoffset              = p_lgl(default = FALSE, tags = "train"),
+        cvoffsetnfolds        = p_int(1L, default = 10L, tags = "train"),
+        return.x              = p_lgl(default = TRUE, tags = "train"),
+        # glmnet::cv.glmnet() parameters
+        lambda                = p_uty(default = NULL, tags = "train"),
+        grouped               = p_lgl(default = TRUE, tags = "train"),
+        trace.it              = p_int(0, 1, default = 0, tags = "train"), # alias: itrace
+        # glmnet::glmnet() parameters
+        type.logistic         = p_fct(c("Newton", "modified.Newton"), default = "Newton", tags = "train"),
+        # prioritylasso:::predict.prioritylasso() parameters
         include.allintercepts = p_lgl(default = FALSE, tags = "predict"),
-        use.blocks = p_uty(default = "all", tags = "predict"),
-
-        # params from cv.glmnet
-        alignment            = p_fct(c("lambda", "fraction"), default = "lambda", tags = "train"),
-        alpha                = p_dbl(0, 1, default = 1, tags = "train"),
-        big                  = p_dbl(default = 9.9e35, tags = "train"),
-        devmax               = p_dbl(0, 1, default = 0.999, tags = "train"),
-        dfmax                = p_int(0L, tags = "train"),
-        eps                  = p_dbl(0, 1, default = 1.0e-6, tags = "train"),
-        epsnr                = p_dbl(0, 1, default = 1.0e-8, tags = "train"),
-        exclude              = p_uty(tags = "train"),
-        exmx                 = p_dbl(default = 250.0, tags = "train"),
-        fdev                 = p_dbl(0, 1, default = 1.0e-5, tags = "train"),
-        gamma                = p_uty(tags = "train"),
-        grouped              = p_lgl(default = TRUE, tags = "train"),
-        intercept            = p_lgl(default = TRUE, tags = "train"),
-        keep                 = p_lgl(default = FALSE, tags = "train"),
-        lambda               = p_uty(tags = "train"),
-        lambda.min.ratio     = p_dbl(0, 1, tags = "train"),
-        lower.limits         = p_uty(default = -Inf, tags = "train"),
-        maxit                = p_int(1L, default = 100000L, tags = "train"),
-        mnlam                = p_int(1L, default = 5L, tags = "train"),
-        mxit                 = p_int(1L, default = 100L, tags = "train"),
-        mxitnr               = p_int(1L, default = 25L, tags = "train"),
-        nlambda              = p_int(1L, default = 100L, tags = "train"),
-        offset               = p_uty(default = NULL, tags = "train"),
-        parallel             = p_lgl(default = FALSE, tags = "train"),
-        penalty.factor       = p_uty(tags = "train"),
-        pmax                 = p_int(0L, tags = "train"),
-        pmin                 = p_dbl(0, 1, default = 1.0e-9, tags = "train"),
-        prec                 = p_dbl(default = 1e-10, tags = "train"),
-        predict.gamma        = p_dbl(default = "gamma.1se", special_vals = list("gamma.1se", "gamma.min"), tags = "predict"), #nolint
-        relax                = p_lgl(default = FALSE, tags = "train"),
-        s                    = p_dbl(0, 1, special_vals = list("lambda.1se", "lambda.min"), default = "lambda.1se", tags = "predict"), #nolint
-        standardize.response = p_lgl(default = FALSE, tags = "train"),
-        thresh               = p_dbl(0, default = 1e-07, tags = "train"),
-        trace.it             = p_int(0, 1, default = 0, tags = "train"),
-        type.gaussian        = p_fct(c("covariance", "naive"), tags = "train"),
-        type.logistic        = p_fct(c("Newton", "modified.Newton"), default = "Newton", tags = "train"),
-        type.multinomial     = p_fct(c("ungrouped", "grouped"), default = "ungrouped", tags = "train"),
-        upper.limits         = p_uty(default = Inf, tags = "train")
+        use.blocks            = p_uty(default = "all", tags = "predict"),
+        # Custom mlr3 parameters
+        adaptive.order        = p_lgl(default = FALSE, tags = "train")
       )
 
       super$initialize(
@@ -100,34 +107,62 @@ LearnerClassifPriorityLasso = R6Class("LearnerClassifPriorityLasso",
       if (is.null(self$model)) {
         stopf("No model stored")
       }
-      coefs = self$model$coefficients
+
+      # uses prioritylasso:::coef.prioritylasso()
+      coefs = stats::coef(self$model)$coefficients
       coefs = coefs[coefs != 0]
       names(coefs)
     }
   ),
+
   private = list(
     .train = function(task) {
-      # get parameters for training
-      pars = self$param_set$get_values(tags = "train")
-      pars$family = "binomial"
-      pars$weights = private$.get_weights(task)
+      pv = self$param_set$get_values(tags = "train")
+      pv$family = "binomial"
+      pv$weights = private$.get_weights(task)
 
-      data = as_numeric_matrix(task$data(cols = task$feature_names))
+      data = as.matrix(task$data(cols = task$feature_names))
       target = task$truth()
-      invoke(prioritylasso::prioritylasso,
-             X = data, Y = target,
-             .args = pars)
-    },
-    .predict = function(task) {
-      newdata = as_numeric_matrix(ordered_features(task, self))
-      pv = self$param_set$get_values(tags = "predict")
-      pv = rename(pv, "predict.gamma", "gamma")
 
-      p = invoke(predict, self$model,
-        newdata = newdata, type = "response",
-        .args = pv)
-      p = drop(p)
-      classnames = self$model$glmnet.fit[[1L]]$classnames
+      # If adaptive.order is TRUE, compute block order and penalty factors from the data
+      res = NULL
+      if (length(pv$blocks) >= 2L && isTRUE(pv$adaptive.order)) {
+        pv$adaptive.order = NULL
+        res = adaptive_block_order(data, target, pv)
+        pv = res$pv
+      }
+
+      model = invoke(
+        prioritylasso::prioritylasso,
+        X = data,
+        Y = target,
+        .args = pv
+      )
+
+      # add block penalty factors to the model object
+      if (!is.null(res)) {
+        model$block.penalty.factors = res$penalty.factors
+      }
+
+      model
+    },
+
+    .predict = function(task) {
+      newdata = as.matrix(ordered_features(task, self))
+      pv = self$param_set$get_values(tags = "predict")
+
+      p = as.numeric(
+        invoke(
+          predict,
+          self$model,
+          newdata = newdata,
+          type = "response",
+          .args = pv
+        )
+      )
+
+      classnames = get_prioritylasso_classnames(self$model, task)
+
       if (self$predict_type == "response") {
         response = fifelse(p <= 0.5, classnames[1L], classnames[2L])
         list(response = drop(response))